pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "illuma-samc"
version = "0.3.0"
description = "Production-quality Stochastic Approximation Monte Carlo (SAMC) for PyTorch"
readme = "README.md"
license = {text = "Non-Commercial Research Use Only"}
requires-python = ">=3.11"
authors = [
    {name = "Frank Shih", email = "fshih37@gmail.com"},
]
keywords = ["mcmc", "sampling", "bayesian", "pytorch", "samc", "multimodal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "torch>=2.0",
]

[project.urls]
Homepage = "https://github.com/FrankShih0807/illuma-samc"
Repository = "https://github.com/FrankShih0807/illuma-samc"
"Bug Tracker" = "https://github.com/FrankShih0807/illuma-samc/issues"

[project.optional-dependencies]
viz = ["matplotlib>=3.7"]
progress = ["tqdm>=4.66"]
full = ["matplotlib>=3.7", "tqdm>=4.66"]
scripts = ["pyyaml>=6.0"]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "matplotlib>=3.7",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["reference/sample_code.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
